# sandaga-backend/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sandaga App
  description: >
    'This is a ***Sandaga app API***.'
  termsOfService: https://github.com/maximan81/sandaga-backend/blob/master/LICENSE
  contact:
    email: info@sandaga.com
  license:
    name: MIT
    url: https://github.com/maximan81/sandaga-backend/blob/master/LICENSE
  version: 1.0.0
...
servers:
  - url: https://sandaga.swagger.io/v2
...
paths:
...
  /api/v1/subcategories:
    get:
      tags:
        - subcategories
      summary: Returns all the matched subcategories
      description: Returns the subcategories that matches the given query criteria
      operationId: querySubcategories
      responses:
        200:
          description: For successful fetch.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subcategory'
    post:
      tags:
        - subcategories
      summary: Add a new subcategory
      description: Creates a new subcategory
      operationId: addSubcategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSubcategoryReq'
      responses:
        201:
          description: For successful user creation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subcategory'
  /api/v1/subcategories/{id}:
    get:
      tags:
        - subcategories
      summary: Returns a subcategory
      description: Returns the subcategory that matches the given subcategory ID
      operationId: getSubCategory
      parameters:
        - name: id
          in: path
          description: Subcategory Identifier
          required: true
          schema:
            type: string
      responses:
        200:
          description: For successful fetch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subcategory'
    delete:
      tags:
        - subcategories
      summary: Deletes subcategory
      description: Deletes subcategory based on given card ID.
      operationId: deleteSubcategoryById
      parameters:
        - name: id
          in: path
          description: subcategory Identifier
          required: true
          schema:
            type: string
      responses:
        202:
          description: Accepts the deletion request and perform deletion. If ID does not exist, does nothing.
          content: { }
    put:
      tags:
        - subcategories
      summary: Replace a subcategory
      description: Replace a subcategory with given id.
      operationId: replaceSubcategoryById
      parameters:
        - name: id
          in: path
          description: subcategory Identifier
          required: true
          schema:
            type: string
      requestBody:
        description: Subcategory object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subcategory'
      responses:
        204:
          description: Subcategory update successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subcategory'
        404:
          description: Given subcategory ID doesn't exist
          content: { }
...
components:
  schemas:
...
    Subcategory:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        name:
          type: string
        channel:
          type: string
      xml:
        name: Subcategory
    AddSubcategoryReq:
      description: Request body for subcategory
      type: object
      properties:
        label:
          description: label of the Subcategory
          type: string
        name:
          description: name of the Subcategory
          type: string
        channel:
          description: channel of the Subcategory
          type: string
        category_id:
          description: category id of the Category
          type: string
...